+++ R/manytrade-memberships.R
@@ -3,12 +3,13 @@
 #' @format The memberships database is a list that contains the
 #' following 2 datasets: GPTAD_MEM, DESTA_MEM.
 #' For more information and references to each of the datasets used,
-#' please use the `data_source()`, `data_contrast()`, and `data_evolution()`
-#' functions.
+#' please use the `data_source()` and `data_contrast()` functions.
 #'\describe{
 #' \item{GPTAD_MEM: }{A dataset with 2221 observations and the following
-#' 6 variables: Country_ID, Country, Title, Beg, Signature, Force.}
+#' 9 variables: many_ID, GPTAD_ID, Country_ID, Country, Title, Beg, Signature, 
+#' Force, treaty_ID.}
 #' \item{DESTA_MEM: }{A dataset with 7772 observations and the following
-#' 6 variables: DESTA_ID, Country, Title, Beg, Signature, Force.}
+#' 8 variables: many_ID, DESTA_ID, Country, Title, Beg, Signature, Force, 
+#' treaty_ID.}
 #' }
  "memberships"