+++ R/manytrade-references.R
@@ -1,0 +1,11 @@
+#' references database documentation
+#'
+#' @format The references database is a list that contains the
+#' following 1 datasets: DESTA_REF.
+#' For more information and references to each of the datasets used,
+#' please use the `data_source()` and `data_contrast()` functions.
+#'\describe{
+#' \item{DESTA_REF: }{A dataset with 1000 observations and the following
+#' 3 variables: many_ID1, RefType, many_ID2.}
+#' }
+ "references"