+++ data-raw/memberships/DESTA_MEM/prepare-DESTA_MEM.R
@@ -1,7 +1,7 @@
 # DESTA_MEM Preparation Script
 
 # This is a template for importing, cleaning, and exporting data
-# ready for the qPackage.
+# ready for the many universe.
 
 # Stage one: Collecting data
 DESTA_MEM <- readxl::read_excel("data-raw/memberships/DESTA_MEM/DESTA.xlsx")
@@ -11,23 +11,39 @@ DESTA_MEM <- readxl::read_excel("data-raw/memberships/DESTA_MEM/DESTA.xlsx")
 # formats of the 'DESTA_MEM' object until the object created
 # below (in stage three) passes all the tests.
 DESTA_MEM <- as_tibble(DESTA_MEM) %>%
-  tidyr::pivot_longer(c("c1":"c91"), names_to = "Member", values_to = "Country", values_drop_na = TRUE) %>%
+  tidyr::pivot_longer(c("c1":"c91"), names_to = "Member", values_to = "Country", 
+                      values_drop_na = TRUE) %>%
   #arrange columns containing countries into one column, with each country in rows corresponding to the treaty it is party to
-  qData::transmutate(DESTA_ID = `base_treaty`,
-                     Title = qCreate::standardise_titles(name),
-                     Signature = qCreate::standardise_dates(as.character(year)),
-                     Force = qCreate::standardise_dates(as.character(entryforceyear))) %>%
+  manydata::transmutate(DESTA_ID = `base_treaty`,
+                     Title = manypkgs::standardise_titles(name),
+                     Signature = manypkgs::standardise_dates(as.character(year)),
+                     Force = manypkgs::standardise_dates(as.character(entryforceyear))) %>%
   dplyr::mutate(Beg = dplyr::coalesce(Signature, Force)) %>%
-  dplyr::select(DESTA_ID, Country, Title, Beg, Signature, Force) %>% #match ISO to country name
+  dplyr::select(DESTA_ID, Country, Title, Beg, Signature, Force) %>%
+  #match ISO to country name
   dplyr::arrange(Beg)
-# qCreate includes several functions that should help cleaning
+
+#Add a treaty_ID column
+DESTA_MEM$treaty_ID <- manypkgs::code_agreements(DESTA_MEM, DESTA_MEM$Title, 
+                                                 DESTA_MEM$Beg)
+
+# Add many_ID column
+many_ID <- manypkgs::condense_agreements(manytrade::agreements, 
+                                         var = c(DESTA$treaty_ID, GPTAD$treaty_ID,
+                                                 LABPTA$treaty_ID, TREND$treaty_ID))
+DESTA_MEM <- dplyr::left_join(DESTA_MEM, many_ID, by = "treaty_ID")
+
+# Re-order the columns
+DESTA_MEM <- dplyr::relocate(DESTA_MEM, many_ID)
+
+# manypkgs includes several functions that should help cleaning
 # and standardising your data.
 # Please see the vignettes or website for more details.
 
 # Stage three: Connecting data
 # Next run the following line to make DESTA_MEM available
-# within the qPackage.
-qCreate::export_data(DESTA_MEM, database = "memberships",
+# within the many universe.
+manypkgs::export_data(DESTA_MEM, database = "memberships",
                      URL = "https://www.designoftradeagreements.org/downloads/")
 # This function also does two additional things.
 # First, it creates a set of tests for this object to ensure adherence
@@ -41,17 +57,6 @@ qCreate::export_data(DESTA_MEM, database = "memberships",
 # present in the data_raw folder of the package for citation purposes.
 # Therefore, please make sure that you have permission to use the dataset
 # that you're including in the package.
-# # This function also does two additional things.
-# First, it creates a set of tests for this object to ensure adherence
-# to certain standards.You can hit Cmd-Shift-T (Mac) or Ctrl-Shift-T (Windows)
-# to run these tests locally at any point.
-# Any test failures should be pretty self-explanatory and may require
-# you to return to stage two and further clean, standardise, or wrangle
-# your data into the expected format.
-# Second, it also creates a documentation file for you to fill in.
-# Please note that the export_data() function requires a .bib file to be
-# present in the data_raw folder of the package for citation purposes.
-# Please make sure that you have permission to use the dataset.
 # To add a template of .bib file to package,
-# run `qCreate::add_bib(memberships, DESTA_MEM)`.
+# run `manypkgs::add_bib(memberships, DESTA_MEM)`.