+++ data-raw/memberships/GPTAD_MEM/prepare-GPTAD_MEM.R
@@ -1,7 +1,7 @@
 # GPTAD_MEM Preparation Script
 
 # This is a template for importing, cleaning, and exporting data
-# ready for the qPackage.
+# ready for the many universe.
 
 # Stage one: Collecting data 
 GPTAD_MEM <- read.csv("data-raw/memberships/GPTAD_MEM/GPTAD.csv")
@@ -11,7 +11,9 @@ GPTAD_MEM <- read.csv("data-raw/memberships/GPTAD_MEM/GPTAD.csv")
 # formats of the 'GPTAD_MEM' object until the object created
 # below (in stage three) passes all the tests.
 GPTAD_MEM <- as_tibble(GPTAD_MEM) %>%
-  dplyr::mutate(Country = gsub("\\\\r\\\\n", "", Membership)) %>% #remove \r\n line break entries
+  dplyr::mutate(GPTAD_ID = dplyr::row_number()) %>%
+  dplyr::mutate(Country = gsub("\\\\r\\\\n", "", Membership)) %>%
+  #remove \r\n line break entries
   dplyr::mutate(Country = stringr::str_replace(Country, "China (Taiwan), Nicaragua", "Taiwan, Nicaragua")) %>%
   #standardize delimitation of values by commas
   dplyr::mutate(Country = stringr::str_replace(Country, "St. Kitts-Nevis-Anguilla", "Saint Kitts and Nevis, Anguilla")) %>%
@@ -30,34 +32,60 @@ GPTAD_MEM <- as_tibble(GPTAD_MEM) %>%
   dplyr::mutate(Country = stringr::str_replace(Country, "Suriname and Trinidad and Tobago", "Suriname, Trinidad and Tobago")) %>%
   dplyr::mutate(Country = stringr::str_replace(Country, "Norway and Switzerland", "Norway, Switzerland")) %>%
   #remove 'and' between country names
-  tidyr::separate_rows(Country, sep=",") %>% #separate column into rows by each country
+  tidyr::separate_rows(Country, sep=",") %>%
+  #separate column into rows by each country
   dplyr::mutate(Country = gsub("^and | and$", "", Country)) %>%
-  dplyr::mutate(Country = ifelse(Country == "", NA, Country)) %>% #standardize empty rows and 
-  dplyr::mutate(Country = ifelse(Country == " ", NA, Country)) %>% #rows with only whitespace to NA
+  dplyr::mutate(Country = ifelse(Country == "", NA, Country)) %>%
+  #standardize empty rows and 
+  dplyr::mutate(Country = ifelse(Country == " ", NA, Country)) %>%
+  #rows with only whitespace to NA
   dplyr::filter(Country != "NA") %>% #remove missing rows
-  dplyr::mutate(Country = trimws(Country, which = c("both", "left", "right"), whitespace = "[ \t\r\n]")) %>% #remove whitespace
+  dplyr::mutate(Country = trimws(Country, which = c("both", "left", "right"), 
+                                 whitespace = "[ \t\r\n]")) %>%
+  #remove whitespace
   dplyr::mutate(Country = ifelse(Country == "", NA, Country)) %>%
   dplyr::filter(Country != "NA",
-                Country != "British") %>% #remove empty rows and redundant 'British' in data
-  dplyr::mutate(Country = qTrade::code_countryname(Country)) %>% #translate French country names and correct spelling
-  dplyr::mutate(Country = dplyr::recode(Country, "EC" = "European Community")) %>% #not included in regex list because of overlaps with other country names
-  dplyr::mutate(Country_ID = countrycode::countrycode(Country, origin = 'country.name', destination = 'iso3n')) %>% #add iso code for country names
-  dplyr::mutate(`Date.of.Signature` = ifelse(`Date.of.Signature`=="n/a", NA, `Date.of.Signature`)) %>%
+                Country != "British") %>%
+  #remove empty rows and redundant 'British' in data
+  dplyr::mutate(Country = manytrade::code_countryname(Country)) %>%
+  #translate French country names and correct spelling
+  dplyr::mutate(Country = dplyr::recode(Country, "EC" = "European Community")) %>%
+  #not included in regex list because of overlaps with other country names
+  dplyr::mutate(Country_ID = countrycode::countrycode(Country, 
+                                                      origin = 'country.name',
+                                                      destination = 'iso3n')) %>%
+  #add iso code for country names
+  dplyr::mutate(`Date.of.Signature` = ifelse(`Date.of.Signature`=="n/a", 
+                                             NA, `Date.of.Signature`)) %>%
   dplyr::mutate(`Date.of.Entry.into.Force` = ifelse(`Date.of.Entry.into.Force`=="N/A", NA, `Date.of.Entry.into.Force`)) %>%
-  qData::transmutate(Title = qCreate::standardise_titles(`Common.Name`),
-                     Signature = qCreate::standardise_dates(`Date.of.Signature`),
-                     Force = qCreate::standardise_dates(`Date.of.Entry.into.Force`)) %>%
+  manydata::transmutate(Title = manypkgs::standardise_titles(`Common.Name`),
+                     Signature = manypkgs::standardise_dates(`Date.of.Signature`),
+                     Force = manypkgs::standardise_dates(`Date.of.Entry.into.Force`)) %>%
   dplyr::mutate(Beg = dplyr::coalesce(Signature, Force)) %>%
-  dplyr::select(Country_ID, Country, Title, Beg, Signature, Force) %>% 
+  dplyr::select(GPTAD_ID, Country_ID, Country, Title, Beg, Signature, Force) %>%
   dplyr::arrange(Beg)
-# qCreate includes several functions that should help cleaning
+
+#Add treaty_ID column
+GPTAD_MEM$treaty_ID <- manypkgs::code_agreements(GPTAD_MEM, GPTAD_MEM$Title, 
+                                                 GPTAD_MEM$Beg)
+
+# Add many_ID column
+many_ID <- manypkgs::condense_agreements(manytrade::agreements,
+                                         var = c(DESTA$treaty_ID, GPTAD$treaty_ID,
+                                                 LABPTA$treaty_ID, TREND$treaty_ID))
+GPTAD_MEM <- dplyr::left_join(GPTAD_MEM, many_ID, by = "treaty_ID")
+
+# Re-order the columns
+GPTAD_MEM <- dplyr::relocate(GPTAD_MEM, many_ID)
+
+# manypkgs includes several functions that should help cleaning
 # and standardising your data.
 # Please see the vignettes or website for more details.
 
 # Stage three: Connecting data
-# Next run the following line to make GPTAD_MEM available
-# within the qPackage.
-qCreate::export_data(GPTAD_MEM, database = "memberships",
+# Next run the following line to make GPTAD_MEM available 
+# within the many universe.
+manypkgs::export_data(GPTAD_MEM, database = "memberships",
                      URL = "https://wits.worldbank.org/gptad/library.aspx")
 # This function also does two additional things.
 # First, it creates a set of tests for this object to ensure adherence